requirements.txt
yfinance
python-dotenv
pydantic
requests
pytest
langgraph        
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
feedparser
fastapi
uvicorn[standard]
streamlit
supervisor